re-organize project
