+++ src/App.jsx
@@ -1,8 +1,8 @@
 import { useState } from 'react'
 import './App.css'
-import TodoList from './TodoList'
-import TodoForm from './TodoForm'
-import TodoListItem from './TodoListItem'
+import TodoList from './features/TodoList/TodoList'
+import TodoForm from './features/TodoForm'
+import TodoListItem from './features/TodoList/TodoListItem'
 
 function App() {
   const [todoList, setTodoList] = useState([])
